Migrated binaries to bin/ directory from tests/ Fixing characterization of logging messages from INFO to WORDY and backwards. Updating documentation to work with new setup and new planner

+++ trunk/p2core/pel/pel_vm.C
@@ -996,7 +996,7 @@ DEF_OP(OR) {
   stackPush(Val_Int64::mk(v1 || v2));
 }
 DEF_OP(RAND) {
-  int32_t r = random();
+  int64_t r = random();
   stackPush(Val_Int64::mk(r));
 }
 DEF_OP(COIN) {
